Improve modularity by splitting OS operations that are not coupled with Docker
